% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example dataset}
\format{
A data frame with 500 rows and 5 variables:
\describe{
\item{sweight}{Survey weight}
\item{cluster}{Cluster ID}
\item{wiq}{Wealth quintile}
\item{wic}{Wealth index continuous score}
\item{stunt5}{Child is stunted (1=yes, 0=no)}
}
}
\source{
Sampled from Bangladesh's 2022 DHS survey
}
\usage{
example_data
}
\description{
A data frame containing individual level survey data used in the examples of the Slope Index of Inequality and the Concentration Index functions.
Data represents children under 5-years
}
\keyword{datasets}
