% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data2.R
\docType{data}
\name{example_data2}
\alias{example_data2}
\title{Example dataset 2}
\format{
A data frame with 38 rows and 3 variables:
\describe{
\item{year}{Year of the survey}
\item{r}{Prevalence of stunting for each admin-1 unit for each year}
\item{r_mean}{Prevalence of stunting (at national level) for each year}
}
}
\source{
Provided by the ICEH Retriever
}
\usage{
example_data2
}
\description{
A data frame containing admin-1 level estimates from 3 Lao's surveys (2006, 2011 and 2017) used in the example of the Mean Absolute Difference
}
\keyword{datasets}
