% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiplot.R
\name{iceh_palette}
\alias{iceh_palette}
\title{ICEH Adaptive Color Palettes}
\usage{
iceh_palette(type = c("wealth", "educ", "area", "viridis"), n = NULL)
}
\arguments{
\item{type}{character: "wealth", "educ", "area", or "viridis".}

\item{n}{integer, optional. Number of colors to return. If \code{NULL}, the base
palette is returned.}
}
\value{
A character vector of hexadecimal color codes.
If \code{n} is \code{NULL}, the function returns the base palette corresponding
to the selected \code{type}. If \code{n} is specified, the function returns
\code{n} interpolated colors generated from the base palette using
\code{grDevices::colorRampPalette()}.
}
\description{
ICEH Adaptive Color Palettes
}
\details{
When \code{type = "viridis"}, the palette is generated using
\code{viridisLite::viridis()}. The package \code{viridisLite} must be installed
to use this option.
}
