% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{package}
\name{ICEHmeasures}
\alias{ICEHmeasures-package}
\alias{ICEHmeasures}
\title{ICEHmeasures: The equiplot graph and complex inequality measures}
\description{
The ICEHmeasures package provides a function to calculate the equiplot, an iconic dot-plot graph for visualizing inequalities, as well as
three complex inequality measures: the slope index of inequality, the concentration index and the mean absolute difference to the mean.
}
\note{
For issues, please visit https://equidade.org/contact
}
\section{Core functions}{

\describe{
\item{\code{equiplot()}}{Creates the equiplot graph for visualizing inequalities across settings, subgroups, time points, among others.}
\item{\code{siilogit()}}{Calculates the slope index for inequality, an absolute measure of inequality expressed as the difference between the extremes of the ranking variable}
\item{\code{cixr()}}{Calculates the relative concentration index, a relative measure of inequality expressed as the cumulative concentration of the outcome across the ranking variable distribution}
\item{\code{mad()}}{Calculates the mean absolute difference from a reference group (often the mean), expressed as the average absolute distance from each subgroup to the reference}
\item{\code{iceh_palette_show()}}{Displays the palette colors.}
}
}

\author{
\strong{Maintainer}: Leonardo Ferreira \email{lferreira@equidade.org}

Authors:
\itemize{
  \item Luisa Arroyave
}

}
\keyword{internal}
